commented text-file format register command
An error is coming from line 346, where the code calls 'register'. This is there so that if the code is pasted into a text window in blender it can be executed and registered with the program. If you're trying to import the script in through the proper installation means you need to comment that line out, as line 342 will evaluate 'true' and call register itself, so it doesn't need to be called again (and produces an error if called). I have changed it so that by default this line (346) is commented out. Uncomment it when using the script via the text window.